Function pulse::swap_scheduler

pub fn swap_scheduler(sched: Box<Scheduler>) -> Option<Box<Scheduler>>

Replace the current Scheduler with your own supplied scheduler. all wait() commands will be run through this scheduler now.

This will return the current TLS scheduler, which may be useful to restore it later.