Struct pulse::Barrier

pub struct Barrier { /* fields omitted */ }

A Barrier can listen for 1 or more Signals. It will only transition to a Pulsed state once all the Signals have Pulsed.

Methods

impl Barrier

pub fn new(pulses: &[Signal]) -> Barrier

Create a new Barrier from an Vector of Siganls