Trait pulse::Scheduler

pub trait Scheduler: Debug {
    fn wait(&self, signal: Signal) -> Result<(), WaitError>;

    fn wait_timeout_ms(
        &self, 
        signal: Signal, 
        timeout: u32
    ) -> Result<(), TimeoutError>;
}

This is the hook into the async wait methods provided by pulse. It is required for the user to override the current system scheduler.

Required Methods

fn wait(&self, signal: Signal) -> Result<(), WaitError>

Wait until the signal is made ready or errored

fn wait_timeout_ms(
&self,
signal: Signal,
timeout: u32
) -> Result<(), TimeoutError>

Wait until the signal is made ready or errored or the timeout has been reached.