Trait pulse::Signals

pub trait Signals {
    fn signal(&self) -> Signal;

    fn wait(&self) -> Result<(), WaitError> { ... }

    fn wait_timeout_ms(&self, ms: u32) -> Result<(), TimeoutError> { ... }
}

allows an object to assert a wait signal

Required Methods

fn signal(&self) -> Signal

Get a signal from a object

Provided Methods

fn wait(&self) -> Result<(), WaitError>

Block the current thread until the object assets a pulse.

fn wait_timeout_ms(&self, ms: u32) -> Result<(), TimeoutError>

Block the current thread until the object assets a pulse. Or until the timeout has been asserted.